mexc_futures_bot/requirements.txt
pandas==2.2.3
python-telegram-bot==20.8
requests==2.32.3
